+++ src/subscribe.rs
@@ -2,27 +2,16 @@
 mod tests;
 
 use std::sync::LazyLock;
-use std::{
-    collections::HashMap,
-    fs::{File, OpenOptions},
-    io::{BufReader, BufWriter},
-    net::SocketAddr,
-    path::{Path, PathBuf},
-    process::exit,
-    sync::Arc,
-};
+use std::{collections::HashMap, net::SocketAddr, path::PathBuf, process::exit, sync::Arc};
 
-use anyhow::{bail, Context, Error, Result};
+use anyhow::{bail, Result};
 use async_channel::{Receiver, Sender};
 use chrono::{TimeZone, Utc};
 use giganto_client::{
     connection::client_handshake,
     frame::{send_raw, RecvError},
     ingest::{
-        network::{
-            Bootp, Conn, DceRpc, Dhcp, Dns, Ftp, Http, Kerberos, Ldap, Mqtt, Nfs, Ntlm, Rdp, Smb,
-            Smtp, Ssh, Tls,
-        },
+        network::{Conn, Dns, Http, Rdp},
         receive_ack_timestamp, send_record_header,
     },
     publish::{
@@ -34,7 +23,7 @@ use giganto_client::{
 };
 use num_traits::ToPrimitive;
 use quinn::{Connection, ConnectionError, Endpoint, RecvStream, SendStream, VarInt, WriteError};
-use serde_json::Value;
+use review_protocol::types::{SamplingKind, SamplingPolicy};
 use tokio::{
     sync::{Mutex, Notify, RwLock},
     task,
@@ -42,101 +31,71 @@ use tokio::{
 };
 use tracing::{error, info, trace, warn};
 
+use crate::time_series::{delete_last_timestamp, write_last_timestamp, SamplingPolicyExt};
 use crate::{
     client::{self, Certs, SERVER_RETRY_INTERVAL},
-    model::{convert_policy, time_series, Period, TimeSeries},
-    request::{RequestedKind, RequestedPolicy},
+    time_series::TimeSeries,
 };
 
 const REQUIRED_GIGANTO_VERSION: &str = "0.23.0";
 const TIME_SERIES_CHANNEL_SIZE: usize = 1;
 const LAST_TIME_SERIES_TIMESTAMP_CHANNEL_SIZE: usize = 1;
-const SECOND_TO_NANO: i64 = 1_000_000_000;
 
 // A hashmap for data transfer to an already created asynchronous task
 pub static INGEST_CHANNEL: LazyLock<RwLock<HashMap<String, Sender<TimeSeries>>>> =
     LazyLock::new(|| RwLock::new(HashMap::new()));
 
-// A hashmap for last series timestamp
-pub static LAST_TRANSFER_TIME: LazyLock<RwLock<HashMap<String, i64>>> =
-    LazyLock::new(|| RwLock::new(HashMap::new()));
+trait FromExt<T> {
+    fn from_ext(t: T) -> Self;
+}
 
-impl From<RequestedKind> for RequestStreamRecord {
-    fn from(k: RequestedKind) -> Self {
+impl FromExt<SamplingKind> for RequestStreamRecord {
+    fn from_ext(k: SamplingKind) -> Self {
         match k {
-            RequestedKind::Conn => Self::Conn,
-            RequestedKind::Dns => Self::Dns,
-            RequestedKind::Http => Self::Http,
-            RequestedKind::Rdp => Self::Rdp,
-            RequestedKind::Smtp => Self::Smtp,
-            RequestedKind::Ntlm => Self::Ntlm,
-            RequestedKind::Kerberos => Self::Kerberos,
-            RequestedKind::Ssh => Self::Ssh,
-            RequestedKind::DceRpc => Self::DceRpc,
-            RequestedKind::Ftp => Self::Ftp,
-            RequestedKind::Mqtt => Self::Mqtt,
-            RequestedKind::Ldap => Self::Ldap,
-            RequestedKind::Tls => Self::Tls,
-            RequestedKind::Smb => Self::Smb,
-            RequestedKind::Nfs => Self::Nfs,
-            RequestedKind::Bootp => Self::Bootp,
-            RequestedKind::Dhcp => Self::Dhcp,
+            SamplingKind::Conn => Self::Conn,
+            SamplingKind::Dns => Self::Dns,
+            SamplingKind::Http => Self::Http,
+            SamplingKind::Rdp => Self::Rdp,
         }
     }
 }
 
+#[allow(clippy::large_enum_variant)]
 pub(crate) enum Event {
     Conn(Conn),
     Dns(Dns),
     Http(Http),
     Rdp(Rdp),
-    Smtp(Smtp),
-    Ntlm(Ntlm),
-    Kerberos(Kerberos),
-    Ssh(Ssh),
-    DceRpc(DceRpc),
-    Ftp(Ftp),
-    Mqtt(Mqtt),
-    Ldap(Ldap),
-    Tls(Tls),
-    Smb(Smb),
-    Nfs(Nfs),
-    Bootp(Bootp),
-    Dhcp(Dhcp),
 }
 
 impl Event {
-    pub(crate) fn column_value(&self, column: usize) -> f64 {
+    pub(crate) fn column_value(&self, column: u32) -> f64 {
         match self {
             Self::Conn(evt) => evt.column_value(column),
             Self::Dns(evt) => evt.column_value(column),
             Self::Http(evt) => evt.column_value(column),
             Self::Rdp(evt) => evt.column_value(column),
-            Self::Smtp(evt) => evt.column_value(column),
-            Self::Ntlm(evt) => evt.column_value(column),
-            Self::Kerberos(evt) => evt.column_value(column),
-            Self::Ssh(evt) => evt.column_value(column),
-            Self::DceRpc(evt) => evt.column_value(column),
-            Self::Ftp(evt) => evt.column_value(column),
-            Self::Mqtt(evt) => evt.column_value(column),
-            Self::Ldap(evt) => evt.column_value(column),
-            Self::Tls(evt) => evt.column_value(column),
-            Self::Smb(evt) => evt.column_value(column),
-            Self::Nfs(evt) => evt.column_value(column),
-            Self::Bootp(evt) => evt.column_value(column),
-            Self::Dhcp(evt) => evt.column_value(column),
         }
+    }
+
+    fn try_new(k: SamplingKind, raw_event: &[u8]) -> Result<Self> {
+        Ok(match k {
+            SamplingKind::Conn => Self::Conn(bincode::deserialize::<Conn>(raw_event)?),
+            SamplingKind::Dns => Self::Dns(bincode::deserialize::<Dns>(raw_event)?),
+            SamplingKind::Http => Self::Http(bincode::deserialize::<Http>(raw_event)?),
+            SamplingKind::Rdp => Self::Rdp(bincode::deserialize::<Rdp>(raw_event)?),
+        })
     }
 }
 
 trait ColumnValue {
-    fn column_value(&self, _: usize) -> f64 {
+    fn column_value(&self, _: u32) -> f64 {
         1_f64
     }
 }
 
 impl ColumnValue for Conn {
-    fn column_value(&self, column: usize) -> f64 {
+    fn column_value(&self, column: u32) -> f64 {
         match column {
             5 => self.duration.to_f64().unwrap_or_default(),
             7 => self.orig_bytes.to_f64().unwrap_or_default(),
@@ -154,38 +113,12 @@ impl ColumnValue for Rdp {}
 
 impl ColumnValue for Http {}
 
-impl ColumnValue for Smtp {}
-
-impl ColumnValue for Ntlm {}
-
-impl ColumnValue for Kerberos {}
-
-impl ColumnValue for Ssh {}
-
-impl ColumnValue for DceRpc {}
-
-impl ColumnValue for Ftp {}
-
-impl ColumnValue for Mqtt {}
-
-impl ColumnValue for Ldap {}
-
-impl ColumnValue for Tls {}
-
-impl ColumnValue for Smb {}
-
-impl ColumnValue for Nfs {}
-
-impl ColumnValue for Bootp {}
-
-impl ColumnValue for Dhcp {}
-
 pub struct Client {
     ingest_addr: SocketAddr,
     publish_addr: SocketAddr,
     server_name: String,
     endpoint: Endpoint,
-    request_recv: Receiver<RequestedPolicy>,
+    request_recv: Receiver<SamplingPolicy>,
     last_series_time_path: PathBuf,
 }
 
@@ -197,7 +130,7 @@ impl Client {
         server_name: String,
         last_series_time_path: PathBuf,
         certs: &Certs,
-        request_recv: Receiver<RequestedPolicy>,
+        request_recv: Receiver<SamplingPolicy>,
     ) -> Self {
         let endpoint =
             client::config(certs).expect("Server configuration error with cert, key or ca_certs");
@@ -213,7 +146,7 @@ impl Client {
 
     pub async fn run(
         self,
-        active_policy_list: Arc<RwLock<HashMap<u32, RequestedPolicy>>>,
+        active_policy_list: Arc<RwLock<HashMap<u32, SamplingPolicy>>>,
         delete_policy_ids: Arc<RwLock<Vec<u32>>>,
         shutdown: Arc<Notify>,
     ) -> Result<()> {
@@ -334,8 +267,8 @@ async fn publish_connection_control(
     server_name: &str,
     endpoint: &Endpoint,
     version: &str,
-    request_recv: &Receiver<RequestedPolicy>,
-    active_policy_list: Arc<RwLock<HashMap<u32, RequestedPolicy>>>,
+    request_recv: &Receiver<SamplingPolicy>,
+    active_policy_list: Arc<RwLock<HashMap<u32, SamplingPolicy>>>,
     delete_policy_ids: Arc<RwLock<Vec<u32>>>,
     last_series_time_path: PathBuf,
 ) -> Result<()> {
@@ -344,11 +277,11 @@ async fn publish_connection_control(
         match publish_connect(endpoint, server_addr, server_name, version).await {
             Ok((conn, send)) => {
                 let req_send = Arc::new(Mutex::new(send));
-                for req_pol in active_policy_list.read().await.values() {
+                for policy in active_policy_list.read().await.values() {
                     if let Err(e) = process_network_stream(
                         conn.clone(),
                         series_send.clone(),
-                        req_pol.clone(),
+                        policy.clone(),
                         req_send.clone(),
                         connection_notify.clone(),
                         active_policy_list.clone(),
@@ -390,11 +323,11 @@ async fn publish_connection_control(
                             sleep(Duration::from_secs(SERVER_RETRY_INTERVAL)).await;
                             continue 'connection;
                         }
-                        Ok(req_pol) = request_recv.recv() => {
+                        Ok(policy) = request_recv.recv() => {
                             if let Err(e) = process_network_stream(
                                 conn.clone(),
                                 series_send.clone(),
-                                req_pol,
+                                policy,
                                 req_send.clone(),
                                 connection_notify.clone(),
                                 active_policy_list.clone(),
@@ -475,42 +408,28 @@ async fn publish_connect(
     Ok((conn, send))
 }
 
-async fn calculate_series_start_time(req_pol: &RequestedPolicy) -> Result<i64> {
-    let mut start: i64 = 0;
-    if let Some(last_time) = LAST_TRANSFER_TIME.read().await.get(&req_pol.id.to_string()) {
-        let period = i64::from(u32::from(Period::from(req_pol.period)));
-        let Some(period_nano) = period.checked_mul(SECOND_TO_NANO) else {
-            bail!("Failed to convert period to nanoseconds");
-        };
-        if let Some(last_timestamp) = last_time.checked_add(period_nano) {
-            start = last_timestamp;
-        }
-    }
-    Ok(start)
-}
-
 #[allow(clippy::too_many_arguments)]
 async fn process_network_stream(
     conn: Connection,
     sender: Sender<TimeSeries>,
-    req_pol: RequestedPolicy,
+    policy: SamplingPolicy,
     send: Arc<Mutex<SendStream>>,
     connection_notify: Arc<Notify>,
-    active_policy_list: Arc<RwLock<HashMap<u32, RequestedPolicy>>>,
+    active_policy_list: Arc<RwLock<HashMap<u32, SamplingPolicy>>>,
     delete_policy_ids: Arc<RwLock<Vec<u32>>>,
     last_series_time_path: PathBuf,
 ) -> Result<()> {
-    let start = calculate_series_start_time(&req_pol).await?;
+    let start = policy.start_timestamp().await?;
     let req_msg = RequestTimeSeriesGeneratorStream {
         start,
-        id: req_pol.id.to_string(),
-        src_ip: req_pol.src_ip,
-        dst_ip: req_pol.dst_ip,
-        sensor: req_pol.node.clone(),
+        id: policy.id.to_string(),
+        src_ip: policy.src_ip,
+        dst_ip: policy.dst_ip,
+        sensor: policy.node.clone(),
     };
     send_stream_request(
         &mut (*send.lock().await),
-        RequestStreamRecord::from(req_pol.kind),
+        RequestStreamRecord::from_ext(policy.kind),
         NodeType::TimeSeriesGenerator,
         req_msg,
     )
@@ -529,17 +448,12 @@ async fn process_network_stream(
     Ok(())
 }
 
-#[inline]
-fn error_in_ts(e: &Error) {
-    error!("Failure in generating time series: {}", e);
-}
-
 #[allow(clippy::too_many_lines)]
 async fn receiver(
     conn: Connection,
     sender: Sender<TimeSeries>,
     connection_notify: Arc<Notify>,
-    active_policy_list: Arc<RwLock<HashMap<u32, RequestedPolicy>>>,
+    active_policy_list: Arc<RwLock<HashMap<u32, SamplingPolicy>>>,
     delete_policy_ids: Arc<RwLock<Vec<u32>>>,
     last_series_time_path: PathBuf,
 ) -> Result<()> {
@@ -550,14 +464,13 @@ async fn receiver(
             bail!("Failed to receive stream id value");
         };
 
-        let req_pol = if let Some(req_pol) = active_policy_list.read().await.get(&id) {
-            req_pol.clone()
+        let policy = if let Some(policy) = active_policy_list.read().await.get(&id) {
+            policy.clone()
         } else {
             bail!("Failed to get runtime policy data");
         };
 
-        let start = calculate_series_start_time(&req_pol).await?;
-        let (policy, mut series) = convert_policy(start, req_pol);
+        let mut series = TimeSeries::try_new(&policy).await?;
 
         info!("Stream's policy : {:?}", policy);
         loop {
@@ -572,254 +485,16 @@ async fn receiver(
             }
             if let Ok((raw_event, timestamp)) = receive_time_series_generator_data(&mut recv).await
             {
-                match policy.kind {
-                    RequestStreamRecord::Conn => {
-                        let (time, Ok(event)) = (
-                            Utc.timestamp_nanos(timestamp),
-                            bincode::deserialize::<Conn>(&raw_event),
-                        ) else {
-                            continue;
-                        };
-                        if let Err(e) =
-                            time_series(&policy, &mut series, time, &Event::Conn(event), &sender)
-                                .await
-                        {
-                            error_in_ts(&e);
-                        }
-                    }
-                    RequestStreamRecord::Dns => {
-                        let (time, Ok(event)) = (
-                            Utc.timestamp_nanos(timestamp),
-                            bincode::deserialize::<Dns>(&raw_event),
-                        ) else {
-                            continue;
-                        };
-                        if let Err(e) =
-                            time_series(&policy, &mut series, time, &Event::Dns(event), &sender)
-                                .await
-                        {
-                            error_in_ts(&e);
-                        }
-                    }
-                    RequestStreamRecord::Rdp => {
-                        let (time, Ok(event)) = (
-                            Utc.timestamp_nanos(timestamp),
-                            bincode::deserialize::<Rdp>(&raw_event),
-                        ) else {
-                            continue;
-                        };
-                        if let Err(e) =
-                            time_series(&policy, &mut series, time, &Event::Rdp(event), &sender)
-                                .await
-                        {
-                            error_in_ts(&e);
-                        }
-                    }
-                    RequestStreamRecord::Http => {
-                        let (time, Ok(event)) = (
-                            Utc.timestamp_nanos(timestamp),
-                            bincode::deserialize::<Http>(&raw_event),
-                        ) else {
-                            continue;
-                        };
-                        if let Err(e) =
-                            time_series(&policy, &mut series, time, &Event::Http(event), &sender)
-                                .await
-                        {
-                            error_in_ts(&e);
-                        }
-                    }
-                    RequestStreamRecord::Smtp => {
-                        let (time, Ok(event)) = (
-                            Utc.timestamp_nanos(timestamp),
-                            bincode::deserialize::<Smtp>(&raw_event),
-                        ) else {
-                            continue;
-                        };
-                        if let Err(e) =
-                            time_series(&policy, &mut series, time, &Event::Smtp(event), &sender)
-                                .await
-                        {
-                            error_in_ts(&e);
-                        }
-                    }
-                    RequestStreamRecord::Ntlm => {
-                        let (time, Ok(event)) = (
-                            Utc.timestamp_nanos(timestamp),
-                            bincode::deserialize::<Ntlm>(&raw_event),
-                        ) else {
-                            continue;
-                        };
-                        if let Err(e) =
-                            time_series(&policy, &mut series, time, &Event::Ntlm(event), &sender)
-                                .await
-                        {
-                            error_in_ts(&e);
-                        }
-                    }
-                    RequestStreamRecord::Kerberos => {
-                        let (time, Ok(event)) = (
-                            Utc.timestamp_nanos(timestamp),
-                            bincode::deserialize::<Kerberos>(&raw_event),
-                        ) else {
-                            continue;
-                        };
-                        if let Err(e) = time_series(
-                            &policy,
-                            &mut series,
-                            time,
-                            &Event::Kerberos(event),
-                            &sender,
-                        )
-                        .await
-                        {
-                            error_in_ts(&e);
-                        }
-                    }
-                    RequestStreamRecord::Ssh => {
-                        let (time, Ok(event)) = (
-                            Utc.timestamp_nanos(timestamp),
-                            bincode::deserialize::<Ssh>(&raw_event),
-                        ) else {
-                            continue;
-                        };
-                        if let Err(e) =
-                            time_series(&policy, &mut series, time, &Event::Ssh(event), &sender)
-                                .await
-                        {
-                            error_in_ts(&e);
-                        }
-                    }
-                    RequestStreamRecord::DceRpc => {
-                        let (time, Ok(event)) = (
-                            Utc.timestamp_nanos(timestamp),
-                            bincode::deserialize::<DceRpc>(&raw_event),
-                        ) else {
-                            continue;
-                        };
-                        if let Err(e) =
-                            time_series(&policy, &mut series, time, &Event::DceRpc(event), &sender)
-                                .await
-                        {
-                            error_in_ts(&e);
-                        }
-                    }
-                    RequestStreamRecord::Ftp => {
-                        let (time, Ok(event)) = (
-                            Utc.timestamp_nanos(timestamp),
-                            bincode::deserialize::<Ftp>(&raw_event),
-                        ) else {
-                            continue;
-                        };
-                        if let Err(e) =
-                            time_series(&policy, &mut series, time, &Event::Ftp(event), &sender)
-                                .await
-                        {
-                            error_in_ts(&e);
-                        }
-                    }
-                    RequestStreamRecord::Mqtt => {
-                        let (time, Ok(event)) = (
-                            Utc.timestamp_nanos(timestamp),
-                            bincode::deserialize::<Mqtt>(&raw_event),
-                        ) else {
-                            continue;
-                        };
-                        if let Err(e) =
-                            time_series(&policy, &mut series, time, &Event::Mqtt(event), &sender)
-                                .await
-                        {
-                            error_in_ts(&e);
-                        }
-                    }
-                    RequestStreamRecord::Ldap => {
-                        let (time, Ok(event)) = (
-                            Utc.timestamp_nanos(timestamp),
-                            bincode::deserialize::<Ldap>(&raw_event),
-                        ) else {
-                            continue;
-                        };
-                        if let Err(e) =
-                            time_series(&policy, &mut series, time, &Event::Ldap(event), &sender)
-                                .await
-                        {
-                            error_in_ts(&e);
-                        }
-                    }
-                    RequestStreamRecord::Tls => {
-                        let (time, Ok(event)) = (
-                            Utc.timestamp_nanos(timestamp),
-                            bincode::deserialize::<Tls>(&raw_event),
-                        ) else {
-                            continue;
-                        };
-                        if let Err(e) =
-                            time_series(&policy, &mut series, time, &Event::Tls(event), &sender)
-                                .await
-                        {
-                            error_in_ts(&e);
-                        }
-                    }
-                    RequestStreamRecord::Smb => {
-                        let (time, Ok(event)) = (
-                            Utc.timestamp_nanos(timestamp),
-                            bincode::deserialize::<Smb>(&raw_event),
-                        ) else {
-                            continue;
-                        };
-                        if let Err(e) =
-                            time_series(&policy, &mut series, time, &Event::Smb(event), &sender)
-                                .await
-                        {
-                            error_in_ts(&e);
-                        }
-                    }
-                    RequestStreamRecord::Nfs => {
-                        let (time, Ok(event)) = (
-                            Utc.timestamp_nanos(timestamp),
-                            bincode::deserialize::<Nfs>(&raw_event),
-                        ) else {
-                            continue;
-                        };
-                        if let Err(e) =
-                            time_series(&policy, &mut series, time, &Event::Nfs(event), &sender)
-                                .await
-                        {
-                            error_in_ts(&e);
-                        }
-                    }
-
-                    RequestStreamRecord::Bootp => {
-                        let (time, Ok(event)) = (
-                            Utc.timestamp_nanos(timestamp),
-                            bincode::deserialize::<Bootp>(&raw_event),
-                        ) else {
-                            continue;
-                        };
-                        if let Err(e) =
-                            time_series(&policy, &mut series, time, &Event::Bootp(event), &sender)
-                                .await
-                        {
-                            error_in_ts(&e);
-                        }
-                    }
-
-                    RequestStreamRecord::Dhcp => {
-                        let (time, Ok(event)) = (
-                            Utc.timestamp_nanos(timestamp),
-                            bincode::deserialize::<Dhcp>(&raw_event),
-                        ) else {
-                            continue;
-                        };
-                        if let Err(e) =
-                            time_series(&policy, &mut series, time, &Event::Dhcp(event), &sender)
-                                .await
-                        {
-                            error_in_ts(&e);
-                        }
-                    }
-                    // pcap is not used in crusher
-                    _ => {}
+                let time = Utc.timestamp_nanos(timestamp);
+                let Ok(event) = Event::try_new(policy.kind, &raw_event) else {
+                    error!(
+                        "Failed to deserialize raw_event for sampling kind: {:?}",
+                        policy.kind
+                    );
+                    continue;
+                };
+                if let Err(e) = series.fill(&policy, time, &event, &sender).await {
+                    error!("Failed to generate time series: {}", e);
                 }
             } else {
                 connection_notify.notify_one();
@@ -906,58 +581,5 @@ async fn receive_time_series_timestamp(
             }
         }
     }
-    Ok(())
-}
-
-async fn write_last_timestamp(
-    last_series_time_path: PathBuf,
-    time_receiver: Receiver<(String, i64)>,
-) -> Result<()> {
-    while let Ok((id, timestamp)) = time_receiver.recv().await {
-        LAST_TRANSFER_TIME.write().await.insert(id, timestamp);
-        let file = OpenOptions::new()
-            .create(true)
-            .write(true)
-            .truncate(true)
-            .open(&last_series_time_path)
-            .context("Failed to open last time series timestamp file")?;
-        serde_json::to_writer(&file, &(*LAST_TRANSFER_TIME.read().await))
-            .context("Failed to write last time series timestamp file")?;
-    }
-    Ok(())
-}
-
-pub async fn read_last_timestamp(last_series_time_path: &Path) -> Result<()> {
-    if last_series_time_path.exists() {
-        let file = File::open(last_series_time_path)
-            .context("Failed to open last time series timestamp file")?;
-        let json: serde_json::Value = serde_json::from_reader(BufReader::new(file))?;
-        let Value::Object(map_data) = json else {
-            bail!("Failed to parse json data, invalid json format");
-        };
-        for (key, val) in map_data {
-            let Value::Number(value) = val else {
-                bail!("Failed to parse timestamp data, invalid json format");
-            };
-            #[rustfmt::skip] // rust-lang/rustfmt#4914
-            let Some(timestamp) = value.as_i64() else {
-                bail!("Failed to convert timestamp data, invalid time data");
-            };
-            LAST_TRANSFER_TIME.write().await.insert(key, timestamp);
-        }
-    }
-    Ok(())
-}
-
-fn delete_last_timestamp(last_series_time_path: &Path, id: u32) -> Result<()> {
-    let file = File::open(last_series_time_path)?;
-    let id = format!("{id}");
-    let mut json: serde_json::Value = serde_json::from_reader(BufReader::new(file))?;
-    if let Value::Object(ref mut map_data) = json {
-        map_data.remove(&id);
-    }
-    let file = File::create(last_series_time_path)?;
-    serde_json::to_writer(BufWriter::new(file), &json)?;
-
     Ok(())
 }